refactor: del SQL/Entities
