+++ src/Interfaces/RoleI.ts
@@ -1,0 +1,9 @@
+/**
+ * Роли пользователей на сайте
+ */
+ export enum RoleT {
+    /** Администратор */
+    administrator = 'administrator',
+    /** Базовый модератор */
+    baseModerator = 'base_moderator',
+}
